+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - [Feature: Add disposable metadata transfer support](https://github.com/Tencent/spring-cloud-tencent/pull/458)
 - [loadBalancer module add unit test.](https://github.com/Tencent/spring-cloud-tencent/pull/468)
 - [docs:update mvnw.](https://github.com/Tencent/spring-cloud-tencent/pull/475)
+- [docs:update configuration metadata.](https://github.com/Tencent/spring-cloud-tencent/pull/473)